# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CTM Application api Documentation",
        "description": "L5 Swagger OpenApi description of CTM api",
        "contact": {
            "email": "tarikulislamnahid15@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/employee/shift/all/filtered": {
            "post": {
                "tags": [
                    "EMPLOYEE-SHIFT"
                ],
                "summary": "get paginated employee shift from database",
                "description": "get paginated employee shift from database",
                "operationId": "getAllEmpShiftPaginated",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "searchText": {
                                        "description": "search text for searching by employee shift name",
                                        "type": "text"
                                    },
                                    "perPage": {
                                        "description": "number of employee shift per page",
                                        "type": "text"
                                    },
                                    "page": {
                                        "description": "page number",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/employee/shift/insert": {
            "post": {
                "tags": [
                    "EMPLOYEE-SHIFT"
                ],
                "summary": "insert a Employee Shift with Shift Details",
                "operationId": "InsertEmpShift",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Shift name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Shift name"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "shift Type is : regular OR scheduled"
                                    },
                                    "start_at": {
                                        "type": "string",
                                        "example": "if type is regular start_at get value else null"
                                    },
                                    "end_at": {
                                        "type": "string",
                                        "example": "if type is regular end_at get value else null"
                                    },
                                    "weekdays": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "item_cost": {
                                                    "type": "integer",
                                                    "example": ""
                                                },
                                                "item_name": {
                                                    "type": "string",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "weekday": "sun",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "mon",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "tue",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "wed",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "thu",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "fri",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 1
                                            },
                                            {
                                                "weekday": "sat",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 1
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/employee/shift/update": {
            "post": {
                "tags": [
                    "EMPLOYEE-SHIFT"
                ],
                "summary": "update a employee shift",
                "description": "update a employee shift",
                "operationId": "UpdateEmpShift",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "Shift id"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Shift name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Shift name"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "shift Type is : regular OR scheduled"
                                    },
                                    "start_at": {
                                        "type": "string",
                                        "example": "if type is regular start_at get value else null"
                                    },
                                    "end_at": {
                                        "type": "string",
                                        "example": "if type is regular end_at get value else null"
                                    },
                                    "weekdays": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "item_cost": {
                                                    "type": "integer",
                                                    "example": ""
                                                },
                                                "item_name": {
                                                    "type": "string",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "weekday": "sun",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "mon",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "tue",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "wed",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "thu",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 0
                                            },
                                            {
                                                "weekday": "fri",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 1
                                            },
                                            {
                                                "weekday": "sat",
                                                "start_at": "02:00",
                                                "end_at": "02:00",
                                                "is_weekend": 1
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful Insert operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/employee/shift/delete": {
            "post": {
                "tags": [
                    "EMPLOYEE-SHIFT"
                ],
                "summary": "delete a employee Shift",
                "description": "delete a employee Shift",
                "operationId": "DeleteEmpShift",
                "requestBody": {
                    "description": "enter inputs",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of the employee shift",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful Insert operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}